Extract AbstractCommand class
As Commands will be capable of locating and executing sub-commands the commands
assume those roles from `Application` and `Application` in turn becomes a _type_ of commands.

This PR extracts the shared functionality to an `AbstractCommand` class.
